Added fully functional RTL elements

+++ src/Objects/Nodes/NodeManager.js
@@ -1,6 +1,7 @@
 import { stateManager, InteractionMode, WorldObject } from "../../State/StateManager";
 import { Node, NodeType } from "./Node";
 import { Wire } from "./Wire";
+import { stateNeg } from "../../logic";
 
 
 /* TODO:
@@ -18,6 +19,7 @@ export class NodeManager {
         this.inputCounter = 0;
         this.nodeCounter = 0;
         this.outputCounter = 0;
+        this.universalNodeCounter = 0;
 
         this.wireHolder = document.createElement("div");
         this.wireHolder.style.width = "1px";
@@ -33,6 +35,15 @@ export class NodeManager {
         this.outputNodeHeight = 15;
 
         stateManager.connectOutputTrigger.subscribe(() => this.connectOutputNode());
+        stateManager.recalcualteResistanceTrigger.subscribe(() => {
+            if(stateManager.transistorPresent.get()) {
+                this.inputNodes.forEach((inputNode) => {
+                    if(!(inputNode.isResistorNode || inputNode.isTransistorNode)) {
+                        this.setupNodePaths(inputNode);
+                    }
+                });
+            }
+        });
     }
 
     getNodeById(nodeId) {
@@ -69,11 +80,12 @@ export class NodeManager {
     }
 
     createInputNode(x, y, mouseX, mouseY, isComponentNode) {
-        const id = `${NodeType.INPUT}-${this.inputCounter}`;
+        const id = `${NodeType.INPUT}-${this.universalNodeCounter}`;
         const node = new Node(this.world, id, NodeType.INPUT, x - this.inputNodeWidth / 2, y - this.inputNodeHeight / 2, 
             this.inputNodeWidth, this.inputNodeHeight, isComponentNode);
         this.inputNodes.push(node);
         this.inputCounter++;
+        this.universalNodeCounter++;
 
         if(!isComponentNode) {
             stateManager.interactionMode.set(InteractionMode.CREATING_NODE);
@@ -93,18 +105,20 @@ export class NodeManager {
     }
 
     createNode(x, y) {
-        const id = `${NodeType.NODE}-${this.nodeCounter}`;
+        const id = `${NodeType.NODE}-${this.universalNodeCounter}`;
         const node = new Node(this.world, id, NodeType.NODE, x - this.nodeWidth / 2, y - this.nodeHeight / 2, 
             this.nodeWidth, this.nodeHeight, false);
         this.nodes.push(node);
 
-        const inputId = stateManager.inputNodeId.get();
-        node.inputNodeId = inputId;
         node.parentNodeId = stateManager.lastNodeId.get();
-
         const parentNode = this.getNodeById(node.parentNodeId);
-        parentNode.childNodeIds.push(node.id);
-        if(parentNode.childNodeIds.length > 1) parentNode.isJoint = true;
+        parentNode.addChildNode(node.id);
+
+        if(parentNode.inputNodeId === -1) {
+            node.inputNodeId = parentNode.id;
+        } else {
+            node.inputNodeId = parentNode.inputNodeId;
+        }
 
         const wire = new Wire(parentNode, node, this.wireHolder);
         node.wires.push(wire);
@@ -119,14 +133,16 @@ export class NodeManager {
         });
 
         this.nodeCounter++;
+        this.universalNodeCounter++;
     }
 
     createOutputNode(x, y, mouseX, mouseY, isComponentNode) {
-        const id = `${NodeType.OUTPUT}-${this.outputCounter}`;
+        const id = `${NodeType.OUTPUT}-${this.universalNodeCounter}`;
         const node = new Node(this.world, id, NodeType.OUTPUT, x - this.outputNodeWidth / 2, y - this.outputNodeHeight / 2, 
             this.outputNodeWidth, this.outputNodeHeight, isComponentNode);
         this.outputNodes.push(node);
         this.outputCounter++;
+        this.universalNodeCounter++;
 
         if(!isComponentNode) {
             stateManager.interactionMode.set(InteractionMode.CREATING_NODE);
@@ -152,21 +168,166 @@ export class NodeManager {
         const outputNode = this.getNodeById(outputNodeId);
         const node = this.getNodeById(stateManager.currentNodeId.get());
 
-        node.childNodeIds.push(outputNodeId);
+        node.addChildNode(outputNodeId)
         const wire = new Wire(outputNode, node, this.wireHolder);
         outputNode.wires.push(wire);
-        outputNode.childNodeIds.push(node.id);
+        outputNode.addChildNode(node.id);
 
         outputNode.connectLogicStates(node.logicState);
         node.logicState.signal();
 
+        let inputNodeId;
         if(node.inputNodeId === -1) {
-            outputNode.inputNodeId = node.id;
+            inputNodeId = node.id;
         } else {
-            outputNode.inputNodeId = node.inputNodeId;
+            inputNodeId = node.inputNodeId;
         }
-        const inputNode = this.getNodeById(outputNode.inputNodeId);
-        inputNode.outputNodeIds.push(outputNodeId);
+
+        const inputNode = this.getNodeById(inputNodeId);
+        if((outputNode.isTransistorNode || outputNode.isResistorNode) && !inputNode.isComponentNode) {
+            this.propagateInputNodeId(outputNode, inputNodeId);
+        } else {
+            outputNode.inputNodeId = inputNodeId;
+            inputNode.outputNodeIds.push(outputNodeId);
+        }
+
+        this.setupNodePaths(inputNode);
+    }
+
+    propagateInputNodeId(currentNode, inputNodeId) {
+        currentNode.inputNodeId = inputNodeId;
+
+        let nextNode;
+        if(currentNode.nodeType === NodeType.OUTPUT && (currentNode.isTransistorNode || currentNode.isResistorNode)) {
+            nextNode = this.getNodeById(currentNode.componentChildNodeId);
+            nextNode.inputNodeId = inputNodeId;
+            this.propagateInputNodeId(nextNode, inputNodeId);
+        } else if(currentNode.nodeType === NodeType.INPUT || currentNode.nodeType === NodeType.NODE) {
+            currentNode.childNodeIds.forEach((childNodeId) => {
+                nextNode = this.getNodeById(childNodeId);
+                this.propagateInputNodeId(nextNode, inputNodeId);
+            });
+        } else {
+            const inputNode = this.getNodeById(inputNodeId);
+            const outputNodeId = currentNode.id;
+            if(!inputNode.outputNodeIds.includes(outputNodeId)) {
+                inputNode.outputNodeIds.push(outputNodeId);
+            }
+        }
+    }
+
+    getPrecedingNode(node) {
+        if(!node) return;
+
+        let nextNode;
+        switch(node.nodeType) {
+            case NodeType.INPUT: {
+                if(node.isComponentNode) {
+                    if(node.isTransistorNode || node.isResistorNode) nextNode = this.getNodeById(node.componentParentNodeId);
+                }
+
+                break;
+            } case NodeType.NODE: {
+                nextNode = this.getNodeById(node.parentNodeId);
+
+                break;
+            } case NodeType.OUTPUT: {
+                if(node.childNodeIds.length !== 0) nextNode = this.getNodeById(node.childNodeIds[0]);
+
+                break;
+            }
+        }
+        return nextNode;
+    }
+
+    /*findNearestJoint(node) {
+        if(!node) {
+            this.resistorCount = 0;
+            this.groundedPath = false;
+            return;
+        }
+
+        let jointNode;
+        let nextNode;
+        if(node.isJoint) {
+            jointNode = node;
+        } else {
+            if(node.nodeType === NodeType.OUTPUT) {
+                if(node.isGrounded) this.groundedPath = true;
+                if(node.isResistorNode) this.resistorCount++;
+            }
+            nextNode = this.getPrecedingNode(node);
+            if(nextNode) if(nextNode.isJoint) jointNode = nextNode;
+        }
+
+        if(jointNode) {
+            node.resistorCount = this.resistorCount;
+            node.connectedToGround = this.groundedPath;
+            this.resistorCount = 0;
+            this.groundedPath = false;
+            return jointNode;
+        } else if(nextNode) {
+            return this.findNearestJoint(nextNode);
+        } else {
+            this.resistorCount = 0;
+            this.groundedPath = false;
+            return;
+        }
+    }*/
+
+    setupNodePaths(inputNode) {
+        // TODO: "Hide" this somewhere?
+        this.inputNodes.forEach((node) => node.isOpen = true);
+        this.nodes.forEach((node) => node.isOpen = true);
+        this.outputNodes.forEach((node) => node.isOpen = true);
+
+        if(stateManager.resistorPresent.get() || stateManager.transistorPresent.get() || stateManager.groundPresent.get()) {
+            const pathResistances = [];
+            let resistance = 0;
+            inputNode.outputNodeIds.forEach((outputNodeId) => {
+                let nextNode = this.getNodeById(outputNodeId);
+                while(nextNode.id !== inputNode.id) {
+                    if(nextNode.isResistorNode && nextNode.nodeType === NodeType.OUTPUT) resistance++;
+                    if(nextNode.isGrounded) resistance += -0.5;
+                    if(nextNode.isTransistorNode && nextNode.nodeType === NodeType.OUTPUT && !nextNode.transistorOn) {
+                        resistance += 1000;
+                    } 
+
+                    nextNode = this.getPrecedingNode(nextNode);
+                }
+                pathResistances.push(resistance);
+                resistance = 0;
+            });
+
+            const minResistance = Math.min(...pathResistances);
+            for(let pathCounter = 0; pathCounter < pathResistances.length; pathCounter++) {
+                if(pathResistances[pathCounter] !== minResistance) {
+                    let node = this.getNodeById(inputNode.outputNodeIds[pathCounter]);
+                    while(node.id !== inputNode.id) {
+                        node.isOpen = false;
+
+                        node = this.getPrecedingNode(node);
+                    }
+                }
+            }
+            for(let pathCounter = 0; pathCounter < pathResistances.length; pathCounter++) {
+                if(pathResistances[pathCounter] === minResistance) {
+                    let node = this.getNodeById(inputNode.outputNodeIds[pathCounter]);
+                    while(node.id !== inputNode.id) {
+                        node.isOpen = true;
+
+                        node = this.getPrecedingNode(node);
+                    }
+                }
+            }
+        }
+
+        this.inputNodes.forEach((node) => {
+            if(!node.isComponentNode) {
+                node.logicState.set(stateNeg(node.logicState));
+                node.logicState.set(stateNeg(node.logicState));
+            }
+        });
     }
 
     deleteNode(node) {
@@ -179,7 +340,7 @@ export class NodeManager {
         node.wires = [];
 
         if(node.unsubFromParentLogicState) node.unsubFromParentLogicState();
-
+        
         node.element.remove();
 
         switch(node.nodeType) {
@@ -199,12 +360,14 @@ export class NodeManager {
     deleteOutputNode(outputNode) {
         if(outputNode.childNodeIds.length > 0) {
             const node = this.getNodeById(outputNode.childNodeIds[0]);
-            node.childNodeIds.splice(node.childNodeIds.indexOf(outputNode.id), 1);
+            node.removeChildNode(outputNode.id);
         }
 
         if(outputNode.inputNodeId !== -1) {
             const inputNode = this.getNodeById(outputNode.inputNodeId);
             inputNode.outputNodeIds.splice(inputNode.outputNodeIds.indexOf(outputNode.id), 1);
+
+            this.setupNodePaths(inputNode);
         }
 
         this.deleteNode(outputNode);
@@ -215,9 +378,13 @@ export class NodeManager {
             const childNode = this.getNodeById(childId);
             if(childNode.nodeType !== NodeType.OUTPUT) {
                 this.deleteChildren(childNode);
-                this.deleteNode(childNode);
+                if(childNode.nodeType !== NodeType.INPUT) this.deleteNode(childNode);
+                else childNode.isOpen = true;
             } else {
                 childNode.childNodeIds = [];
+                
+                const inputNode = this.getNodeById(childNode.inputNodeId);
+                if(inputNode) inputNode.outputNodeIds.splice(inputNode.outputNodeIds.indexOf(childNode.id), 1);
                 childNode.inputNodeId = -1;
                 
                 const wire = childNode.wires[0];
@@ -226,25 +393,29 @@ export class NodeManager {
                 wire.element.remove();
                 childNode.wires = [];
 
-                childNode.unsubFromParentLogicState();
+                childNode.isOpen = true;
                 childNode.logicState.set(0);
+                childNode.unsubFromParentLogicState();
             }
         });
         node.childNodeIds = [];
+        node.isJoint = false;
+
+        if(node.inputNodeId !== -1) {
+            const inputNode = this.getNodeById(node.inputNodeId);
+            if(inputNode) this.setupNodePaths(inputNode);
+        }
     }
 
     deleteNodeChain(node) {
         this.deleteChildren(node);
         if(node.parentNodeId !== -1) {
             const parentNode = this.getNodeById(node.parentNodeId);
-            if(parentNode) {
-                parentNode.childNodeIds.splice(parentNode.childNodeIds.indexOf(node.id), 1);
-                if(parentNode.childNodeIds.length <= 1) parentNode.isJoint = false;
-            }
+            if(parentNode) parentNode.removeChildNode(node.id);
         }
         this.deleteNode(node);
 
-        stateManager.setDefaultState();
+        stateManager.setDefaultInteractionState();
     }
 
     deleteGeneralNode(node) {
@@ -260,7 +431,6 @@ export class NodeManager {
         if(node.isComponentNode) this.deleteChildren(node);
         else if(node.nodeType === NodeType.INPUT && node.childNodeIds.length > 0) {
             this.deleteChildren(node);
-        }
-        else this.deleteGeneralNode(node);
+        } else this.deleteGeneralNode(node);
     }
 }
